+++ libyul/optimiser/SimplificationRules.cpp
@@ -21,15 +21,16 @@
 
 #include <libyul/optimiser/SimplificationRules.h>
 
-#include <libyul/optimiser/ASTCopier.h>
-#include <libyul/optimiser/Semantics.h>
-#include <libyul/optimiser/SyntacticalEquality.h>
-#include <libyul/optimiser/DataFlowAnalyzer.h>
-#include <libyul/backends/evm/EVMDialect.h>
 #include <libyul/AST.h>
 #include <libyul/Utilities.h>
+#include <libyul/backends/evm/EVMDialect.h>
+#include <libyul/optimiser/ASTCopier.h>
+#include <libyul/optimiser/DataFlowAnalyzer.h>
+#include <libyul/optimiser/Semantics.h>
+#include <libyul/optimiser/SyntacticalEquality.h>
 
 #include <libevmasm/RuleList.h>
+#include <libsolutil/StringUtils.h>
 
 using namespace std;
 using namespace solidity;
@@ -249,8 +250,7 @@ Expression Pattern::toExpression(shared_ptr<DebugData const> const& _debugData) 
 		for (auto const& arg: m_arguments)
 			arguments.emplace_back(arg.toExpression(_debugData));
 
-		string name = instructionInfo(m_instruction).name;
-		transform(begin(name), end(name), begin(name), [](auto _c) { return tolower(_c); });
+		string name = util::toLower(instructionInfo(m_instruction).name);
 
 		return FunctionCall{_debugData,
 			Identifier{_debugData, YulString{name}},