+++ test/libevmasm/Optimiser.cpp
@@ -1250,8 +1250,8 @@ BOOST_AUTO_TEST_CASE(jumpdest_removal_subassemblies)
 	// tag unifications (due to block deduplication) is also
 	// visible at the super-assembly.
 
-	Assembly main;
-	AssemblyPointer sub = make_shared<Assembly>();
+	Assembly main{false, {}};
+	AssemblyPointer sub = make_shared<Assembly>(true, string{});
 
 	sub->append(u256(1));
 	auto t1 = sub->newTag();
@@ -1278,7 +1278,6 @@ BOOST_AUTO_TEST_CASE(jumpdest_removal_subassemblies)
 	main.append(u256(8));
 
 	Assembly::OptimiserSettings settings;
-	settings.isCreation = false;
 	settings.runInliner = false;
 	settings.runJumpdestRemover = true;
 	settings.runPeephole = true;
@@ -1287,7 +1286,7 @@ BOOST_AUTO_TEST_CASE(jumpdest_removal_subassemblies)
 	settings.runConstantOptimiser = true;
 	settings.evmVersion = solidity::test::CommonOptions::get().evmVersion();
 	settings.expectedExecutionsPerDeployment = OptimiserSettings{}.expectedExecutionsPerDeployment;
-;
+
 	main.optimise(settings);
 
 	AssemblyItems expectationMain{