+++ test/libsolidity/syntaxTests/inlineAssembly/natspec_memory_safe.sol
@@ -1,0 +1,6 @@
+contract C {
+	function f() public pure {
+		// @solidity memory-safe-assembly
+		assembly {}
+	}
+}